+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Fixed video playback issues on machines that do not support browser OpenGL acceleration
 * Fixed WebGL issues on 10.5 and 10.6 Intel builds
 * Fixed GitHub website issues through built-in Private Elements extension. Thank you to Private Elements developers for their contributions
+* Fixed apple.com gallery issues through built-in DOMMatrix transform list polyfill
 * Fixed various browser UI issues
 * Updated FFmpeg library to 7.1.5
 * Updated NSS security library to 3.90.13 (UXP)